Fix brave/brave-ios#1918: Send in page innerText to ads classifier (brave/brave-ios#1930)
Instead of sending plain HTML, we now send in the document's `innerText`, which doesn't include any HTML tags, images, etc.
This allows the ads classifier to properly classify pages to their correct type

+++ BraveRewardsUI/Tipping/TippingViewController.swift
@@ -175,7 +175,7 @@ class TippingViewController: UIViewController, UIViewControllerTransitioningDele
           // TODO: Handle started tip process
         }
       } else {
-        self.state.ledger.tipPublisherDirectly(self.publisherInfo, amount: Int32(amount), currency: "BAT") { _ in
+        self.state.ledger.tipPublisherDirectly(self.publisherInfo, amount: Double(amount), currency: "BAT") { _ in
           // TODO: Handle started tip process
         }
       }
